Moved pkg_util.py to apio/utils
