+++ apio/managers/installer.py
@@ -12,11 +12,10 @@ from pathlib import Path
 from typing import Tuple
 import shutil
 from click import secho
-from apio import pkg_util
 from apio.apio_context import ApioContext
 from apio.managers.downloader import FileDownloader
 from apio.managers.unpacker import FileUnpacker
-from apio.utils import util
+from apio.utils import util, pkg_util
 
 
 def _construct_package_download_url(
